Moved DTO files to Models directory
